move generate stuff into generate[_test].go
